Set up firebase directly, drop extra dependencies
* take out the assumption that apps will being using embefire &
  ember-data
  * set up ref to firebase directly in initializer rather than through
  adapter
