+++ app/authenticators/firebase.js
