+++ bower.json
@@ -3,8 +3,8 @@
   "dependencies": {
     "handlebars": "2.0.0",
     "jquery": "^1.11.1",
+    "firebase": "~2.2.3",
     "ember": "1.9.1",
-    "ember-data": "1.0.0-beta.12",
     "ember-resolver": "~0.1.11",
     "loader.js": "ember-cli/loader.js#1.0.1",
     "ember-cli-shims": "ember-cli/ember-cli-shims#0.0.3",
@@ -14,7 +14,6 @@
     "ember-qunit-notifications": "0.0.5",
     "qunit": "~1.17.1",
     "ember-simple-auth": "0.7.2",
-    "mockfirebase": "~0.9.4",
-    "emberfire": "~1.3.2"
+    "mockfirebase": "~0.9.4"
   }
 }
