+++ index.js
@@ -50,10 +50,5 @@ module.exports = {
       development: app.bowerDirectory + '/firebase/firebase-debug.js',
       production: app.bowerDirectory + '/firebase/firebase.js'
     });
-
-    this.app.import({
-      development: app.bowerDirectory + '/emberfire/dist/emberfire.js',
-      production: app.bowerDirectory + '/emberfire/dist/emberfire.min.js'
-    });
   }
 };
