+++ package.json
@@ -29,9 +29,7 @@
     "ember-cli-qunit": "0.3.0",
     "ember-cli-simple-auth-testing": "^0.7.2",
     "ember-cli-uglify": "1.0.1",
-    "ember-data": "1.0.0-beta.12",
     "ember-export-application-global": "^1.0.0",
-    "emberfire": "^1.3.2",
     "express": "^4.8.5",
     "glob": "^4.0.5"
   },
